+++ types/kafkajs.d.ts
@@ -8,7 +8,8 @@ import {
   DeleteRecordsResult,
   Node,
   AclOperationTypes,
-  Uuid
+  Uuid,
+  IsolationLevel
 } from './rdkafka'
 
 // Admin API related interfaces, types etc; and Error types are common, so
@@ -22,7 +23,8 @@ export {
   DeleteRecordsResult,
   Node,
   AclOperationTypes,
-  Uuid
+  Uuid,
+  IsolationLevel
 } from './rdkafka'
 
 export interface OauthbearerProviderResponse {
@@ -419,4 +421,9 @@ export type Admin = {
     includeAuthorizedOperations?: boolean,
     timeout?: number
   }): Promise<{ topics: Array<ITopicMetadata> }>
+  fetchTopicOffsets(topic: string,
+    options?: {
+      timeout?: number,
+      isolationLevel: IsolationLevel
+    }): Promise<Array<SeekEntry & { high: string; low: string }>>
 }